+++ extensions/api/data-management/policydefinition/src/test/java/org/eclipse/dataspaceconnector/api/datamanagement/policy/PolicyApiControllerTest.java
@@ -17,11 +17,13 @@ package org.eclipse.dataspaceconnector.api.datamanagement.policy;
 import org.eclipse.dataspaceconnector.api.datamanagement.policy.service.PolicyService;
 import org.eclipse.dataspaceconnector.api.exception.ObjectExistsException;
 import org.eclipse.dataspaceconnector.api.exception.ObjectNotFoundException;
+import org.eclipse.dataspaceconnector.api.query.QuerySpecDto;
 import org.eclipse.dataspaceconnector.api.result.ServiceResult;
+import org.eclipse.dataspaceconnector.api.transformer.DtoTransformerRegistry;
 import org.eclipse.dataspaceconnector.policy.model.Policy;
 import org.eclipse.dataspaceconnector.spi.monitor.Monitor;
-import org.eclipse.dataspaceconnector.spi.query.Criterion;
-import org.eclipse.dataspaceconnector.spi.query.SortOrder;
+import org.eclipse.dataspaceconnector.spi.query.QuerySpec;
+import org.eclipse.dataspaceconnector.spi.result.Result;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
@@ -32,6 +34,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.argThat;
+import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.ArgumentMatchers.isA;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
@@ -39,18 +42,19 @@ import static org.mockito.Mockito.when;
 
 class PolicyApiControllerTest {
 
-    private final PolicyService policyService = mock(PolicyService.class);
+    private final PolicyService service = mock(PolicyService.class);
+    private final DtoTransformerRegistry transformerRegistry = mock(DtoTransformerRegistry.class);
     private PolicyApiController controller;
 
     @BeforeEach
     void setup() {
         var monitor = mock(Monitor.class);
-        controller = new PolicyApiController(monitor, policyService);
+        controller = new PolicyApiController(monitor, service, transformerRegistry);
     }
 
     @Test
     void getPolicyById() {
-        when(policyService.findById("id")).thenReturn(Policy.Builder.newInstance().build());
+        when(service.findById("id")).thenReturn(Policy.Builder.newInstance().build());
 
         var policyDto = controller.getPolicy("id");
 
@@ -59,28 +63,31 @@ class PolicyApiControllerTest {
 
     @Test
     void getPolicyById_notExists() {
-        when(policyService.findById("id")).thenReturn(null);
+        when(service.findById("id")).thenReturn(null);
 
         assertThatThrownBy(() -> controller.getPolicy("id")).isInstanceOf(ObjectNotFoundException.class);
     }
 
     @Test
     void getAllPolicies() {
+        when(service.query(any())).thenReturn(List.of(Policy.Builder.newInstance().build()));
+        when(transformerRegistry.transform(isA(QuerySpecDto.class), eq(QuerySpec.class)))
+                .thenReturn(Result.success(QuerySpec.Builder.newInstance().offset(10).build()));
+        var querySpec = QuerySpecDto.Builder.newInstance().build();
 
-        when(policyService.query(any())).thenReturn(List.of(Policy.Builder.newInstance().build()));
-        var allPolicies = controller.getAllPolicies(1, 10, "field=value", SortOrder.ASC, "field");
-
-        var result = allPolicies.get(0);
-        var size = allPolicies.size();
+        var allPolicies = controller.getAllPolicies(querySpec);
 
         assertThat(allPolicies).hasSize(1);
-        verify(policyService).query(argThat(s ->
-                s.getOffset() == 1 &&
-                        s.getLimit() == 10 &&
-                        s.getFilterExpression().get(0).equals(new Criterion("field", "=", "value")) &&
-                        s.getSortOrder().equals(SortOrder.ASC) &&
-                        s.getSortField().equals("field")
-        ));
+        verify(service).query(argThat(s -> s.getOffset() == 10));
+        verify(transformerRegistry).transform(isA(QuerySpecDto.class), eq(QuerySpec.class));
+    }
+
+    @Test
+    void getAll_throwsExceptionIfQuerySpecTransformFails() {
+        when(transformerRegistry.transform(isA(QuerySpecDto.class), eq(QuerySpec.class)))
+                .thenReturn(Result.failure("Cannot transform"));
+
+        assertThatThrownBy(() -> controller.getAllPolicies(QuerySpecDto.Builder.newInstance().build())).isInstanceOf(IllegalArgumentException.class);
     }
 
     @Test
@@ -99,11 +106,11 @@ class PolicyApiControllerTest {
 
         var policy = Policy.Builder.newInstance().build();
 
-        when(policyService.create(any())).thenReturn(ServiceResult.success(policy));
+        when(service.create(any())).thenReturn(ServiceResult.success(policy));
 
         controller.createPolicy(policyDefinition);
 
-        verify(policyService).create(isA(Policy.class));
+        verify(service).create(isA(Policy.class));
     }
 
     @Test
@@ -121,7 +128,7 @@ class PolicyApiControllerTest {
                 .build();
 
         var policy = Policy.Builder.newInstance().build();
-        when(policyService.create(any())).thenReturn(ServiceResult.conflict("already exists"));
+        when(service.create(any())).thenReturn(ServiceResult.conflict("already exists"));
 
         assertThatThrownBy(() -> controller.createPolicy(policyDefinition)).isInstanceOf(ObjectExistsException.class);
     }
@@ -129,20 +136,20 @@ class PolicyApiControllerTest {
 
     @Test
     void deletePolicy() {
-        when(policyService.deleteById("id")).thenReturn(ServiceResult.success(Policy.Builder.newInstance().build()));
+        when(service.deleteById("id")).thenReturn(ServiceResult.success(Policy.Builder.newInstance().build()));
         controller.deletePolicy("id");
-        verify(policyService).deleteById("id");
+        verify(service).deleteById("id");
     }
 
     @Test
     void deletePolicy_notFound() {
-        when(policyService.deleteById("id")).thenReturn(ServiceResult.notFound("Not found"));
+        when(service.deleteById("id")).thenReturn(ServiceResult.notFound("Not found"));
         assertThatThrownBy(() -> controller.deletePolicy("id")).isInstanceOf(ObjectNotFoundException.class);
     }
 
     @Test
     void deletePolicy_conflicts() {
-        when(policyService.deleteById("id")).thenReturn(ServiceResult.conflict("Conflicting"));
+        when(service.deleteById("id")).thenReturn(ServiceResult.conflict("Conflicting"));
         assertThatThrownBy(() -> controller.deletePolicy("id")).isInstanceOf(ObjectExistsException.class);
     }
 }