+++ extensions/azure/blobstorage/blob-core/src/testFixtures/java/org/eclipse/dataspaceconnector/azure/blob/core/AzureStorageTestFixtures.java
@@ -45,6 +45,10 @@ public class AzureStorageTestFixtures {
         return faker.lorem().characters();
     }
 
+    public static String createSharedAccessSignature() {
+        return faker.lorem().characters();
+    }
+
     private AzureStorageTestFixtures() {
     }