+++ CHANGELOG.md
@@ -24,11 +24,14 @@ in the detailed section referring to by linking pull requests or issues.
 * Check to avoid duplicated module names (#1190)
 * Implement Catalog service for Data Management API (#1195)
 * Add strict body validation for REST endpoints (#1128)
+* Add validation on query endpoints (#1258)
 * Dependency injection using factory/provider methods (#1056)
 * Provisioned resource information in Data Management API (#1221)
 * Add custom Jackson (de)serializer for `XMLGregorianCalendar` (#1226)
 * Add contract validation rule (#1239)
 * Harmonize setting names in `data-plane-transfer` (#1164)
+* Add Blob transfer Architectural Decision Record (#1259)
+* Add component tests coverage to the codecov coverage report (#1246)
 
 #### Changed
 
@@ -55,6 +58,8 @@ in the detailed section referring to by linking pull requests or issues.
 * Break lease after TransferProcessManager status check (#1214)
 * Fix path conflicts between `CatalogApiController` and `FederatedCatalogApiController` (#1225)
 * Always use configured IDS API path in IDS webhook address (#1249)
+* Fix Azure storage transfer (#1245)
+* Throw exception if `IdentityProviderKeyResolver` cannot get keys at startup (#1266)
 
 ## [milestone-3] - 2022-04-08